test(api): separate Quran route coverage
